+++ src/appliers/object/schema.ts
@@ -1,5 +1,5 @@
 import {Key, Values} from "../../libs/types";
-import {SchemaObject, applySchemaToCore} from "../../libs/applySchemaToCore";
+import {SchemaObject, applySchemaObjectCore} from "../../libs/applySchemaObjectCore";
 import {onErrorDefault, onFinishedDefault} from "../../libs/BaseSchema";
 
 export interface Options
@@ -22,6 +22,6 @@ export function applyTo<T>(values: Values, options: Options, keyStack: Key[]): v
 		return false;
 	}
 
-	values.output = applySchemaToCore(values.output, options.schemaObject, onErrorDefault, onFinishedDefault, keyStack);
+	values.output = applySchemaObjectCore(values.output, options.schemaObject, onErrorDefault, onFinishedDefault, keyStack);
 	return false;
 }